fix: CommentService Interface 분리
