+++ src/main/java/com/study/todoparty/controller/CommentController.java
@@ -5,7 +5,7 @@ import com.study.todoparty.dto.requestDto.CreateCommentRequestDto;
 import com.study.todoparty.dto.requestDto.UpdateCommentRequestDto;
 import com.study.todoparty.dto.responseDto.CommentResponseDto;
 import com.study.todoparty.dto.responseDto.CommonResponseDto;
-import com.study.todoparty.service.CommentService;
+import com.study.todoparty.service.CommentServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,7 +16,7 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 @RequestMapping("/todos")
 public class CommentController {
-    private final CommentService commentService;
+    private final CommentServiceImpl commentServiceImpl;
 
     // 댓글 작성 API : 선택한 할일이 있다면 댓글을 등록
     // 조건 : 토큰을 검사하여, 유요한 토큰일 경우에만 댓글 작성 가능
@@ -24,7 +24,7 @@ public class CommentController {
     // 반환 정보 : 등록된 댓글
     @PostMapping("/{todoId}/comments/create")
     public ResponseEntity<CommentResponseDto> createComment(@PathVariable Long todoId, @RequestBody CreateCommentRequestDto request, @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        CommentResponseDto response = commentService.createComment(todoId, request, userDetails.getUser());
+        CommentResponseDto response = commentServiceImpl.createComment(todoId, request, userDetails.getUser());
         return ResponseEntity.status(HttpStatus.CREATED).body(response);
     }
 
@@ -35,7 +35,7 @@ public class CommentController {
 
     @PutMapping("/{todoId}/comments/update")
     public ResponseEntity<CommonResponseDto> updateComment(@PathVariable Long todoId, @RequestBody UpdateCommentRequestDto request, @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        CommentResponseDto response = commentService.updateComment(todoId, request, userDetails.getUser());
+        CommentResponseDto response = commentServiceImpl.updateComment(todoId, request, userDetails.getUser());
         return ResponseEntity.ok().body(response);
     }
 
@@ -45,7 +45,7 @@ public class CommentController {
     // 반환 정보 : 성공 메시지, 상태코드 = ResponseEntity<CommonResponse>
     @DeleteMapping("/{todoId}/comments/delete/{commentId}")
     public ResponseEntity<CommonResponseDto> deleteComment(@PathVariable Long todoId, @PathVariable Long commentId, @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        commentService.deleteComment(todoId, commentId, userDetails.getUser());
+        commentServiceImpl.deleteComment(todoId, commentId, userDetails.getUser());
         return ResponseEntity.ok().body(new CommonResponseDto("댓글이 삭제되었습니다.", HttpStatus.OK.value()));
     }
 }
